Extracted a prettyTime method to DRY up formatting code. Noticed (and took) opportunity to eliminate another local variable in play().
It turns out the that original code had a bug in the prettye printing of times,
showing 10:1 instead of 10:01. Fixed this bug, which changed the expected value
of the characterization test testConsoleOutput. In a prodution setting, this
change would require additional investigation to make sure no client code is
depending on the original (wrong) behavior.